backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
pymongo==4.6.0
python-dotenv==1.0.0
PyJWT==2.10.1
google-auth==2.41.1
requests==2.32.5
gunicorn==23.0.0
